WIP: Update and Modified the existing code to ITK Exteranl module
 - Reorganized the folder structure to ITK External module's structure
 - Modified CMakeLists.txt and added itk-module.cmake

 - Modified itkHighPriorityRealTimeProbe based on new itkResourceExpandedProbe class and
   modified itkResourceProbe2.

 - Updated itkHighPriorityRealTimeProbesCollector with new itkResourceExpandedProbesCollector

 - Add two test codes (itkHighPriorityRealTimeProbeTest.cxx and itkHighPriorityRealTimeProbesCollectorTest3. cxx) for
   itkHighPriorityRealTimeProbe and itkHighPriorityRealTimeProbesCollector,respectively.
