--- Benchmarks/ClockTest/itkHighPriorityRealTimeProbesCollectorTest.cxx
@@ -1,9 +1,0 @@
-#include <iostream>
-
-
-int main()
-{
-  std::cout << "Hello World" << std::endl;
-
-  return 1;
-}
