--- Benchmarks/Filtering/CMakeLists.txt
@@ -1,5 +1,0 @@
-message("Performance test for ITK Filters")
-message("ITK_DATA_ROOT:::" ${ITK_DATA_ROOT})
-
-
-
