--- Benchmarks/Registration/CMakeLists.txt
@@ -1,1 +1,0 @@
-message("Performance test fot ITK Registration")
