--- Benchmarks/Segmentation/CMakeLists.txt
@@ -1,1 +1,0 @@
-message("Performance test for ITK Segmentation")
