+++ CMakeLists.txt
@@ -2,9 +2,9 @@ cmake_minimum_required(VERSION 2.8)
 
 project(ITK_BENCHMARK)
 
-option(${PROJECT_NAME}_SUPERBUILD "" ON)
+option(ITK_BENCHMARK_SUPERBUILD "" ON)
 
-if (${PROJECT_NAME}_SUPERBUILD)
+if (ITK_BENCHMARK_SUPERBUILD)
   # External Project Build of ITK
   include(ExternalProject)
 
@@ -14,7 +14,7 @@ if (${PROJECT_NAME}_SUPERBUILD)
 
   ExternalProject_Add(ITK
     GIT_REPOSITORY "git://itk.org/ITK.git"
-    GIT_TAG ef3e18aa90a64c1913fa29f9e31711666ca679de
+    GIT_TAG b755a25a2fec5ced65ab0a8ca8c2276bba402f67
     UPDATE_COMMAND ""
     DOWNLOAD_DIR ${ITK_SOURCE_DIR}
     SOURCE_DIR ${ITK_SOURCE_DIR}
@@ -42,13 +42,13 @@ if (${PROJECT_NAME}_SUPERBUILD)
     INSTALL_COMMAND ""
     )
 
-  set(ITK_BENCHMARK_EXECUTABLE_OUTPUT_PATH  "${CMAKE_BINARY_DIR}/bin")
+  #set(ITK_BENCHMARK_EXECUTABLE_OUTPUT_PATH  "${CMAKE_BINARY_DIR}/bin")
 
-  ExternalProject_Add(${PROJECT_NAME}
+  ExternalProject_Add(ITK_BENCHMARK
     UPDATE_COMMAND ""
     DOWNLOAD_COMMAND ""
     SOURCE_DIR ${CMAKE_SOURCE_DIR}
-    BINARY_DIR ${CMAKE_BINARY_DIR}#/${PROJECT_NAME}-build
+    BINARY_DIR ${CMAKE_BINARY_DIR}#/ITK_BENCHMARK-build
     LOG_DOWNLOAD 1
     LOG_UPDATE 0
     LOG_CONFIGURE 0
@@ -57,7 +57,7 @@ if (${PROJECT_NAME}_SUPERBUILD)
     LOG_INSTALL 0
     CMAKE_GENERATOR ${gen}
     CMAKE_ARGS
-      -DCMAKE_RUNTIME_OUTPUT_DIRECTORY:STRING=${ITK_BENCHMARK_EXECUTABLE_OUTPUT_PATH}
+      #-DCMAKE_RUNTIME_OUTPUT_DIRECTORY:STRING=${ITK_BENCHMARK_EXECUTABLE_OUTPUT_PATH}
       -DCMAKE_CXX_COMPILER:STRING=${CMAKE_CXX_COMPILER}
       -DCMAKE_CXX_COMPILER_ARG1:STRING=${CMAKE_CXX_COMPILER_ARG1}
       -DCMAKE_C_COMPILER:STRING=${CMAKE_C_COMPILER}
@@ -75,11 +75,11 @@ if (${PROJECT_NAME}_SUPERBUILD)
 endif()
 
 # Find ITK Library
-find_package(ITK REQUIRED)
-include(${ITK_USE_FILE})
+#find_package(ITK REQUIRED)
+#include(${ITK_USE_FILE})
 
 #Enable CTest
-enable_testing()
+#enable_testing()
 
 # Add sub-directory
-add_subdirectory(Benchmarks)
+add_subdirectory(PerformanceBenchmarks)
