+++ PerformanceBenchmarks/CMakeLists.txt
@@ -1,0 +1,12 @@
+cmake_minimum_required(VERSION 2.9)
+project(PerformanceBenchmarks)
+
+set(PerformanceBenchmarks_LIBRARIES PerformanceBenchmarks)
+
+if(NOT ITK_SOURCE_DIR)
+  find_package(ITK REQUIRED)
+  list(APPEND CMAKE_MODULE_PATH ${ITK_CMAKE_DIR})
+  include(ITKModuleExternal)
+else()
+  itk_module_impl()
+endif()
