+++ PerformanceBenchmarks/itk-module.cmake
@@ -1,0 +1,12 @@
+set(DOCUMENTATION "ITK Performance Benchmarks")
+
+itk_module(PerformanceBenchmarks
+  ENABLE_SHARED
+  PRIVATE_DEPENDS
+    ITKCommon
+  TEST_DEPENDS
+    ITKTestKernel
+    ITKCommon
+  DESCRIPTION
+  "${DOCUMENTATION}"
+)
