+++ PerformanceBenchmarks/src/CMakeLists.txt
@@ -1,0 +1,9 @@
+set( PerformanceBenchmarks_SRCS
+itkHighPriorityRealTimeClock.cxx
+itkHighPriorityRealTimeProbe.cxx
+itkHighPriorityRealTimeProbesCollector.cxx
+)
+
+add_library( PerformanceBenchmarks ${ITK_LIBRARY_BUILD_TYPE} ${PerformanceBenchmarks_SRCS})
+itk_module_link_dependencies()
+itk_module_target(PerformanceBenchmarks)
